Use pip-compile's backtracking dependency resolver
The CI pipeline began failing three days ago, due some dependency
changes in an upstream dependency.

It seems the pip-compile step of our tox environments became unable to
resolve a full dependency graph from this point on (related to versions
of pylint, it seems).

Using the backtracking resolver of pip-compile allows it to select older
dependencies as long our pinned requirements are satisfied.