Lock to an old ciscconfparse during testing
ciscoconfparse is a dependency of NAV's direct dependency
napalm. However, the latest few versions of ciscoconfparse have been
packaged with dependencies that are only relevant for developers, not
for production use. These dependencies conflict with our test
environment dependencies.

Until we upgrade to Napalm 4 (which ditches ciscoconfparse), we lock it
to an older, stable version during test runs.