Get rid of coverage database access errors
This should get rid of all the intermittent access errors from
coverage. These are mainly due to the fact that certain subprocesses
during test setup (such as database initialization inside test
container) will run as root or other users that do not have write access
to the source root directory.

This points coverage to use a coverage database file in the reports
directory, and ensures this is world-writeable (!) so all subprocesses
can access it.