Add ugly hack for netsnmp import on MacOS
MacOS' System Integrity Protection wreaks havoc with NAV and its test
suite, due to its dependency on pynetsnmp / NET-SNMP:

pynetsnmp will load libcrypto using ctypes, which aborts the python
process on MacOS with an error message like this:

> WARNING: .tox/integration-py39-django32/bin/python is loading libcrypto
> in an unsafe way

When OpenSSL is installed on MacOS through homebrew, the environment
variable DYLD_LIBRARY_PATH must be set for processes to load the correct
version of libcrypto. However, newer versions of MacOS sanitizes this
environment variable on many levels.

The whole debaucle of trying to fix this in a sane way is described at
https://hynek.me/articles/macos-dyld-env/

This workaround is no prettier, and will only work if libssl is
installed using Homebrew, or if the value of DYLD_LIBRARY_PATH can be
gleaned from LD_LIBRARY_PATH.