Make gunicorn a session fixture
Instead of always start/stopping gunicorn at the start/end of an
integration test session, this makes it a session fixture that tests can
depend on directly.

The new fixture is defined at the top-level `conftest.py` in order for
it to be reusable by any part of the test suite.