Replace create-db.sh with PostgreSQL fixtures
This creates a PostgreSQL fixture that can be depended on by tests that
require a running PostgreSQL server with the NAV database to be present.

The fixture will work with externally set up PostgreSQL servers, like in
the case of GitHub Actions, while it can also provision a PostgreSQL
server with the NAV schema using Docker if run locally.