Make snmpsim fixture work portably
This ensure the snmpsim fixture can run both on Linux and MacOS, by
replacing Linux specific parts (like looking at /proc/net/udp).

This slightly changes the logic of verifying whether the snmpsimd
process is up and running before yielding the fixture value: Instead of
waiting for something to appear to listen to port 1024, this verifies
the presence of the SNMP agent by performing an actual SNMP query.