Never restore stale tox environment caches
The installed dependencies in tox environments are no longer
automatically synced using pip-tools, which means they do not update if
they are stale.

We should therefore never use prefix matching for the cache Action for
the tox environments, or we will end up restoring a stale environment
every time the requirements files change.

This ensures we only restore tox environments when the cache key matches
exactly - meaning that all the tox environments should be rebuilt if any
of the requirements files changed.