Lock pyasn1<0.5.0
`snmpsim` depends on `pyasn1`, but does not lock the dependent
version. `pyasn1` version 0.5.0 was recently released, and it seems
`snmpsim` is incompatible with the new version, causing integration
tests that use snmpsim to fail.

This locks pyasn1 to a version that should still work with the latest
version of snmpsim.