Stop making validation tests for non HTML content
The web crawler would generate HTML validation tests for all reachable
pages, and marked them as *skipped* if the page is blacklisted or the
retrieved page has a content-type that doesn't appear to be
HTML-related.

Skipped tests should be an indicator of some unusual situation or
missing condition that prevents the test from running.  However,
non-HTML pages can never be validated as such, and the *normal*
situation is that these tests are useless.  Since a lot of the crawled
content isn't actually HTML, a large number of SKIP results is produced
in test reports.  These results only serve to hide the results of tests
that were skipped for *factually* unexpected or extraordinary reasons.

This change will filter crawled pages based on blacklist status and
content-type before it the result is used to parametrize the
`test_page_should_be_valid_html` test.