Make signals propagate to remote socat process
This ensures that the remote socat process is torn down when the
forwarding container exits. Before this change, the socat process would
linger forever, causing the forwarding port to stay bound forever on the
remote hop-host.  Any subsequent attempts to restart the forwarding
container would just result in an error that looked like this:

```
socat[10092] E bind(5, {AF=2 0.0.0.0:10000}, 16): Address already in use
```

Using the `-tt` option to ssh forces SSH to allocate a pseudo-terminal,
even if the controlling process isn't attached to a real terminal.
Using the pseudo-terminal ensures that signals are propagated from the
ssh process to the socat process started on the remote, so when the SSH
session dies, socat will also.

We also drop the `-f` option, since that would cause the SSH process to
send the termination signal as soon as it forks and exits.  This uses
the shell to background the ssh process instead.