"Un-parametrize" test_get_navlet_should_return_200
While pytest can accomplish a lot of exciting things, it cannot use
fixtures as input to test parametrization.  While can make a test
depend on a fixture for getting a postgresql instance, the test
discovery phase that generates the tests cannot.

I.e. we cannot get our test parameters from a database unless the
database was already up and running when the test discovery phase
started.  Sad, but true,.

This changes the navlet test to iterate over the available admin navlets
itself.