Fix non-interactive sudo test
`sudo -nv` can be used to verify whether the sudo command is prepared
to execute commands non-interactively, i.e. it will not ask for a
password.

If the password has already been cached by previous sudo commands, or
sudo is configured to give passwordless access, we can move on.
Otherwise, interactive sudo or lack of sudo access cannot be used to
run tests in an automated fashion.