Remove the old, bundled Sphinx bootstrap theme
This theme contains both an old bootstrap version and an ancient jQuery
version.

There should be no need for us to bundle this with NAV at all, as it
exists as a package on PyPI: `sphinx-bootstrap-theme`

The custom.css additions are left in, they need to be moved somewhere
else to remain active.