Pass C_INCLUDE_PATH to test environment
At least two or three of NAVs dependencies may need to be built from
source when pulled in by pip: psycopg2, python-ldap and Pillow.  On
platforms like NixOS, the include files for the necessary C libraries
may not be in standardized FHS locations, so the C_INCLUDE_PATH is used
to convey their locations to the C compiler.