Add toggle for automatic creation of remote users
Add a new flag to the remote user config that by default is off. If
remote users is enabled and the new flag is toggled on, the username in
REMOTE_USER will be used to automatically create a new user with that
username. Automatic creation regardless used to be the default behavior.