Extend SNMPParameters dataclass with SNMPv3 params
A better design might subclass this into v1, v2c and v3 versions, but
inheritance would not be straightforward: community is part of v1 and
v2c, bulk parameters are valid for v2c and v3, and auth and privacy
parameters are only valid for v3.

This bunches everything together, but does not perform any error
detection or corrections.  Incoming profiles are assumed to be
correct and verified as they are inserted into NAV.