Replace read_only and snmp_version Netbox attrs
Netbox shadow objects used to use the pseudo attributes `read_only` and
`snmp_version` to store the SNMP v1/v2c read-only community and SNMP
version.  On the Netbox Django model, these are computed properties,
based on the Netbox' preferred SNMP management profile, and have been
deprecated for a while.

Since ipdevpoll now needs to access more complicated SNMP management
configuration for a Netbox, this changes the Netbox shadow model to
store an SNMPParameters object directly, which will be populated from
the Netbox' preferred SNMP profile.

Usages of the removed attributes have been updated.  Usages mostly
concern themselves with verifying that a device actually has an SNMP
profile at all.

However, the ipdevpoll dataloader used the value of the pseudo
attributes to detect whether a Netbox' management credentials had
changed and needed to be reloaded from the database, and this patch does
not fully replace that.

This more or less fixes #2522