Fix profile fixture
The profile fixture returned a mocked management profile object that
was incomplete.  The new SNMP setup routines inspect the profile
configuration more closely, and would fail.  There is no need to use
a Mock to mock a profile, just create an instance with an actual
configuration dict, but just don't save it anywhere.