Update arnold enable/disable tests
Use the new dependency injection path to insert our mocked SNMP Agent
into change_port_status().

Also, since change_port_status() now gets a partial call to an Snmp
constructor, it will only feed in the IP address of the Netbox, so the
asserts needed to change to account for that.