Handle low-level SnmpTimeoutErrors for SNMPv3
An SNMPv3 conversation may start with a discovery phase, to learn the
remote SNMP entity's authoritative EngineID.  This initial pair of
"discovery" and "report" packages are handled implicitly by NET-SNMP.

However, this can result in a timeout error from the async call to get
a value, rather than a timeout error that we generate "ourselves" when
we have waited too long.  This type of error is raised by pynetsnmp
as "yet another timeout exception class", rather than one of the
existing timeout exceptions.

This adds `SnmpTimeoutError` to the list of handled timeout exceptions
where timeouts are handled explicitly, while in the mibretriever
retrieval methods it is translated to a regular TimeoutError exception.