Properly detect changes to Netbox SNMP profiles
The ipdevpoll-internal Netbox data structure needs to be udpated when
the netbox' preferred SNMP profile has changed in the database, to
ensure we can still communicate with the netbox.

This is simply done by comparing the new computed `snmp_parameters`
attribute.  But, we must also remember to actually store the new value
in the Netbox.copy() method.