Update logic of Netbox.get_preferred_snmp_profile
This changes the logic behind preferred SNMP profile selection, a need
that has become apparent while working with SNMPv3 profiles.

A read-write profile should be acceptable for read operations if no
read-only profile exists.  However, if write access is not required,
but a write profile is available in addition to a read profile, the read
profile should be preferred.

This also updates all callers to the new signature, based on their
actual needs.