Fix SNMP v1/v2c regression in async libraries
The SNMPParameters class expects its SNMP version attribute to be
an integer.  However, some (or perhaps all) SNMPv1/v2c management
profiles tend to store the version numbers as strings, for some reason.

The net result is that the `as_agentproxy_args()` method would fail
to provide a `community` argument to SNMP v1/v2 sessions because it
was unable to compare an integer version number to a string.