Ensure AgentProxy.snmpVersion is prefixed by `v`
If the `snmpVersion` argument to `AgentProxy` is not a version string
prefixed by the letter `v`, `AgentProxy` will silently default to `v2c`.

Using `2c` as a value works, since it defaults then to `v2c`.  Using `3`
also seems to work, as providing further SNMPv3 configuration arguments
to the Net-SNMP library seems to automatically set the session version
to `v3`, regardless of the version parameter.

However, using `1` will not work, as that will also silently default
the session to `v2c`, making SNMPv1 communication impossible.