Do not cycle session id in ensure_account
Avoids session id changing on every request.
session will still be cycled on login by the functions
that directly handle login.

ensure_account either just sets the request.account field
to the match the already logged in user, or sets the account
to be the anonymous user. Neither should trigger a session_id.