Remove redundant build step
The build step used to be there for "caching" purposes, and for the
build files to be stored with owner privileges for the `nav` user,
rather than dropping files owned by root into the mounted `/source`
directory.

Using the `build` module does not give this effect, since it does
everything inside a separate virtual environment, meaning the subsequent
`pip install` command receives no caching benefit.  In fact, we just
end up with two redundant (and potentially slow) builds of NAV and its
dependencies.