Ensure PATH propagates through privilege drops
For the new Docker image to work properly, the PATH variable needs
to propagate through user changes performed by `nav start` commands.
There were, however, two barriers:  First, su is used to drop privileges
before starting nav daemons, but this resets the PATH variable to
the ones defined in `/etc/login.defs`.  Secondly, `nav start` uses
a `-` to tell `su` to use a login shell, in which case the PATH is still
reset.  The login shell option was introduced by a fix for #2218, but
it's not clear that the use or non-use of a login shell was the actual
problem solved, it was just that the order of the arguments needed to
be portable.

Tagged #2218 just in case this turns into another BSD regression.