Test installed scripts instead of internal modules
This replaces the `test_binary_runs` test with `test_script_runs` test.
The former would just enumerate all the modules present in the `nav.bin`
packages, extract their test args and run them directly from that
location.  However, not all the scripts here are compatible with being
invoked from a pip-installed bin-directory shim that calls their main
method directly.

I.e. this updates the test suite to attempt to call all the defined
NAV scripts in the manner they would be run in an installed system,
rather than in an artificial manner.  This should detect additional
problems with the scripts.