Rework PaloAltoArp configuration build/usage
Access to the configuration of PaloAlto API keys is required before
plugin instances are created, so we might as well fetch the global
ipdevpoll config when the plugin class is loaded and cache the details
in the class object itself.

Also, there is no need to inefficiently process the configuration as
a list when a dict will do the trick much faster and with less code.