Move snmpsim to test requirements list
snmpsim belongs inside the tox test environments, and is required for
several tests.  Not sure why it was being installed explicitly in the
Dockerfiles and workflows, when the correct approach is to name it in
the test requirements  (I'm guessing because it is used as a command
and not as a library by the test suite).

This way, it doesn't need to be referenced multiple times.