"Un-parametrize" webcrawler tests
While pytest can accomplish a lot of exciting things, it cannot use
fixtures as input to test parametrization.  While we can make a test
depend on a fixture for getting a running webserver instance, the test
discovery phase that generates the tests cannot.

I.e. we cannot get our test parameters from the webcrawler unless the
web server was already up and running when the test discovery phase
started.  Sad, but true.

This changes the webcrawler into a fixture, and changes the web link
reachability and HTML validation tests to iterate the pages provided
by this session-scoped crawler.

This also considerably shortens the discovery phase, since the crawling
actually takes place during the running of the first test that uses
the fixture.