Start selenium tests by logging in
Instead of creating a "fake" admin session directly in the database,
this changes the overloaded `selenium` fixture to log in to the NAV
interface as an admin, using a username and password, before handing the
selenium driver to the calling test.

It seems as though our recent fixes to avoid session fixation may cause
the whole "session fakery" to become flaky, as the session cookie might
be reset by loading a page, causing the old selenium fixture to time out
waiting for the expected cookie.

There's also the old issue of whether the faked session will be properly
committed to the database and be visible to the web server process that
runs outside of the test session.

These issues can all be circumvented by just simulating a "normal"
login.