Restrict which branch pushes cause a full CI run
Currently, if a PR is made from a repo-local branch, the full test suite
runs twice, because it triggers on both the branch push and the PR
event.

We do not usually keep long-lived feature branches, so it should be ok
to restrict full test runs to the important branches: `master` and any
branch that matches the stable branch pattern.