Don't iterate over dict being modified
This fixes a Python 2-ism that got lost in the conversion to Python 3 a
few years back: The code would iterate over a dictionary while
potentially making modifications to it, and you cannot do that under
Python 3.

This fixes it by instead just iterating over the keys of the original
dictionary that was copied.