Fix incorrect handling of non-existent scopes
The old code seemed to assume that `scope` would be `None` if no
`Prefix` matching the network address could be found in the DB.  But,
of course, Django raises a `DoesNotExist` exception when using
`Prefix.objects.get()` for this purpose.

This changes the logic to catch the exception, but otherwise produce
the originally intended response.