Make base_url fixture a proxy for gunicorn
Most, if not all, the functional tests need a base URL for the web
server and already depend on the base_url fixture.  The URL is no
longer received from the environment, but from the gunicorn fixture,
so this simply makes the base_url fixture a proxy for the gunicorn
fixture.