Fix more broken pping skip logic
Too broad exception catching caused a `NameError` to be masked
and all the tests to be skipped: The `get_root_method` function was
not available at test definition time, since it was defined below the
tests.

This narrows the exception catcher and moves the `get_root_method()`
before the first test is defined.