Use pytest-xvfb
This adds the `pytest-xvfb` plugin as a test requirement and removes
the now redundant tox commands to set up an Xvfb server before the
functional test suite runs.

This is part of wider strategy to make running the test suite more
independent of tox: Tox is there to help us set up test matrices as
separate environments, but the test suite itself should really be able
to run on manually built environments.