Configure NPM for building CSS from Sass using webpack
It is placed in its own directory in order to separate it from the npm setup for frontend tests.
More and more frontend related config things, scripts, code fixes, NAV js modules can potentially be moved here in the process of frontend code modernisation.

NB! I am not sure whether this is a good placement for this dir. Feedback needed