Configure NPM and add webpack config for building CSS from Sass
The essence of this config:
Sass files will be compiled into CSS files into the same folder (/static/css/) as it was done in the original implementation (with libsass and setuptools).

The resulting files in /static/css will have same layout and names as with the previous solution, except for /static/css/foundation folder. It is dropped completely since it was empty when built with setuptools. This is expected since /sass/foundation folder contains only Sass partials.

Named webpack configuration is used in the config. This is to drop the need of having multiple webpack config files per name/target/purpose.