Turn autoescaping off for contents of legend tag in fieldset template
Much simpler solution but it is a code smell and a potential security hole. We don't really have code to back up arbitrary HTML rendering in templates. But also our templates are rendered on backend and served statically, plus legend tag is not an input field (meaning it is not expected to be updated and processed by backend), so it is not the worst place to add |safe. But I still think this should be avaided if possible.