Update PaloaltoArp ipdevpoll plugin to use ManagementProfiles
Prior to this commit, the netboxes handled by the PaloaltoArp
ipdevpoll plugin used to be configured in the `ipdevpoll.conf`
configuration file, but since the netboxes the plugin wants to
handle (i.e. collect Arp information from) already should reside in
the NAV database, this configuration is now instead done through the
SeedDB tool by assigning a HTTP_API ManagementProfile (with `service`
set `Palo Alto ARP` and api_key set to some secret API key) to the
netboxes to be handled.

The prior way to configure the netboxes handled by the PaloAltoArp
plugin implicitly only allowed one API key per netbox (both enforced
in code but also by the configuration syntax). With
ManagementProfiles, it is perfectly possible to assign multiple
profiles (e.g. configurations) of the same type but with different
parameters (e.g. API keys) to the same netbox. Hence the new way to
configure the netboxes allow many API keys per netbox. Thus the
semantics of the plugin must change a little: For any given netbox,
the plugin now assumes there may be multiple API keys, and uses the
ARP results of first API key for which the _do_request method returns
a successful response.

IMPORTANT:
This commit removes the ability to configure the netboxes
handled by the PaloaltoArp plugin the NAV version 5.10 - 5.11 way
through the `[paloaltoarp]` section in the `ipdevpoll.conf`
configuration file.