Allow comparison to other types of Netbox objects
This stub class cannot reasonably expect to always be compared to other
values of the same class.  In fact, most of the time, it will be
compared to instances of the real class that it is stubbing.

Since the comparison is mostly about verifying that the `sysname`
attribute has the same value, this changes the implementation to be
more lenient to avoid stupid crash bugs (and this fixes the broken
regression tests introduced by the parent commit)