Add new functions to build component trails
They way the maintenance UI builds data structures to describe
maintenance components (and their "trails") is quite obtuse, and seems
to be a strange leftover from the olden days before Django was
introduced to NAV.

This can be done a lot cleaner, while improving on how missing
components are described.

These classes and functions represent an alternative way to build
component trails to place in the template render context.