Refactor maintenance edit view function
This replaces `utils.components_for_keys()` with
`get_components_from_keydict()`, which works more like
the newer `get_components()`, but works on POST data
rather than Task object data.

The end result is an edit view that builds the component
trails use in the maintenance in much the same way as the
refactored `view()` view, which means that the HTML template
for the `edit()` view can now also re-use the
`frag-component-trail.html` template.