Update pytest and plugin versions
This updates the version locks of pytests and its plugins used for test
runs.  Mostly due to the fact that our selected combinations of
Python versions, pytest and pytest-cov were generating deprecation
warnings during test runs that did not have anything to do with NAV
itself.

These versions are locked to what are the current latest releases.