Update list limits for alert profiles dropdowns
For some long-forgotten reason, there are limits configured on the
length of dropdown lists in Alert Profiles: One value per match field
is set in the database.  The limits are set to 1000 for all fields, but
some users do in fact have more than 1000 communication rooms registered
in their NAV installations.

The code does not allow unlimited lengths, so a quickfix for now is
to increase the limit ten-fold.  A more long-term fix would be to look
into why there are limitations at all (aside from the fact that
selecting from a dropdown with > 1000 values and no search functionality
is bad UX)