Refactor management handler selection
This pushes the logic of PortAdmin management handler selection to the
handlers themselves, making it easier to implement more complex
selection logic if needed.  It seems more correct to let the handler
classes decide for themselves whether a given Netbox is compatible.

A `can_handle()` class method is added to `ManagementHandler`, whose
default implementation is based on the existing "selection by `VENDOR`
id" scheme.  The extra logic for Cisco-but-not-quite-Cisco can therefore
be pushed into a specific `can_handle()` implementation for the
Cisco handler.