Run test image as root and drop privs instead
Newer versions of gosu will complain loudly about having the setuid bit
set, and refuse to work.  gosu is not really intended to raise the
privilege level of the calling user, but to lower it.

This switches up the use of the test docker image and gosu, running the
container as root by default, but instead using the entrypoint script to
modify the `build` user and then *drop* privileges to that user before
running the container command.