Upgrade test image to Ubuntu 24.04
Ubuntu has changed how its package lists are specified by default.  It
also ships with libsnmp40 rather than libsnmp35.  Finally, the most
important change is that pip no longer is allowed to install Python
packages globally, so tox and its associates are instead installed to a
virtualenv whose bin directory is added to the default PATH of the
container.