Use OS release version as part of cache key
Some Python packages are compiled from source during the build, and
they will be linked with specific versions of system libraries.  Once
the OS release changes (e.g. when GitHub updates what version of
Ubuntu `ubuntu-latest` refers to), some modules in the cache may be
incompatible with the new release, and need to be recompiled.

Such is the case for python-ldap after upgrading to Ubuntu 24.04, for
example.