Send map widget GET request to info-search view
Seemed like the form sent a request to home (/) directly,
when it probably should have been sent to /search/ instead.
This at least makes it so if you search for a room and it finds
a match it opens the page for the room,
and if there is no match then it still shows the search page but
with the message that nothing was found.

If this is how the map widget is supposed to function is
uncertain to me, I dont actually know how it originally worked
before it broke (if it ever worked)