Allow non-unique UID/GID combo for build user
The usermod/groupmod commands used to ensure the build user has the same
UID/GID as the mounted /source directory will fail if the UID or GID
happens to exist already in the Docker image used.  This would case
strange permission errors, since the build user would not be properly
modified then the container command starts.