Lower acceptable coverage reduction
1% of the NAV codebase would be between 400 and 500 lines, which is
actually a very big number.  It is not clear from the CodeCov docs that
floating point numbers are supported here, so we'll just have to commit
this and test if it works.  0.1% should be between 40 and 50 lines. It's
still on the high side, but we can reduce it further if we increase
explicit coverage of randomly exercised code lines.