Define project dependencies in pyproject.toml
We've fully switched to using pyproject.toml for the project definition,
but the dependencies were not declared outside of the existing set of
requirements files.

The requirements files are fine for defining various environments for
testing, but the full dependency list should be declared in the project
metadata.

This more or less copies all the existing requirements into the
appropriate sections of pyproject.toml.  Minimal runtime dependencies,
optional runtime dependencies, and also dependency groups for
non-runtime requirements, like development or testing of the codebase.