fixup! Add endpoint for JWT refresh tokens
Adds some basic validation of the incoming token.
No need to really check the format or anything,
if its not a valid token in any way it will fail
when we check the hash against the database.
The important part is that the incoming token is
a string that can actually be hashed