Require newest setuptools for build
In order to ensure the setuptools used to build NAV properly supports
the License metadata changes specified by PEP-639, this updates the
minimum setuptools version required for the build system to the latest
version.

It also updates the setuptools_scm version to ensure the latest version
is used (which is assumed to still be compatible with the newest
setuptools).