Add email field to Account
As suggested in code review, many libraries expect the Django user
account to have an email field.  The AbstractUser implementation even
provides a method to get the name of the field that contains the user's
email address.

NAV currently has no need for emails attached directly to user objects,
so this field is just being added for the sake of compatibility.  It
may slip into usage at a later stage.