Handle queries from browseable API
request.data seems to be automatically conferted to a list
when sending requests via cURL (and in tests), but when you
use the browseable API request.data is instead a QueryDict
that containst the JSON string.

This adds support for both types, but its abit messy.. I assume
ModelView has something built in that handles this stuff, so maybe
Django has a function that can do this for us? Perhaps a Mixin or
something like that