Lock down API permissions for regular users
Giving regular NAV user accounts full API access is a mistake and a
security vulnerability that must be closed.  This allowed them to
add and change netbox information, create users and even elevate their
own privileges by giving themselves membership of the administrator
group.

This change locks down the API for non-admin users, except for the Alert
endpoint, which is a read-only endpoint that non-privileged users must
have access to in order to use the Status tool of the web front-end.

It also slightly cleans up how permissions are defined, by replacing
hard-to-read code with simple permissions combinations using the `|`
operator.