Add permission handling for JWT tokens
Now requires permissions to be included as claims.

Write rights is declared via the 'write' claim using
bool values. True means the token has write access.

Endpoint permissions are declared via the
'permission' claim, which should be a list
containing the endpoints the token should have acess
to in the same way that the old tokens do