Hardcode Dockerfile arch to amd64
It turns out this Dockerfile cannot be used on a Mac with Apple Silicon:
The Google Chrome debian repository does not published packages for the
`arm64` platform, so the step that installs Google Chrome fails.

Forcing the build platform to linux/amd64 ensures this image is usable
for testing also on a Mac.

Docker will complain about the hardcoded value, but it doesn't seem to
matter much for our use (since this is not a production image that we
publish anywhere).  Also, we hope to get rid of this entire Dockerfile
in the near future, by revamping our test suite dependency handling.

This is analogous to 950d5145e13dead66ed6a5f6d4381f7422a5decd