Use Django authentication in `do_login`
This replaces the authentication mechanism used in the `do_login` view
to instead employ Django's authentication system. The default list
of authentication backends include only the `ModelBackend`, which
will use NAV's `Account` model for authentication.

This temporarily disconnects LDAP authentication, as part of a proof-of-
concept. LDAP authentication will have to be added as a separate
back-end.

Additionally, it turns out the `AUTH_USER_MODULE` was set incorrectly,
since the app label we use for NAV models is simply `models`.