experiment: adjust psa to support NAV locked
While django.contrib.auth controls whether a user may attempt to log in
via a field on the table, `is_active`, NAV uses whether there is an
acceptable password stored in order to calculate the value of `locked`.

If logging in remotely and creating a user on first attempt, normally
that user is password-less, hence locked, hence cannot log in to NAV.
This commit adjusts the psa pipeline to ask the user to set a password
on first log in.

Whether we want to allow the creation of users on first log in is
something we need to decide on, or whether it should be togglable.