Add libldap-common to devcontainer image
This ensures basic OpenLDAP library configuration is installed in the
devcontainer. These configuration files are necessary in order to point
the OpenLDAP library to the correct set of CA certificates to consider
when verifying LDAP server connections. Without this, setting up NAV
to authenticate against an SSL-protected LDAP server will just fail.