Lock devcontainer postgres to version 17
Using the "latest" tag runs the risk of spontaneous upgrades during
environment restarts.  The data directory is persistent, but newer
versions of PostgreSQL cannot use the data directory of an older version
without some specific intervention to upgrade the data directory.

It's best to keep the upgrade behavior predictable; this does it by
locking to a specific PostgreSQL version.

Additionally, this ensures the main devcontainer is provisioned with
PostgreSQL client programs that match the selected PostgreSQL server
version, so that programs like `pg_dump` will work correctly.