Add support for NAV_CONFIG_DIR env var
The NAV_CONFIG_DIR environment variable can be used to insert a custom
directory as the first entry in the list of directories NAV will search
for its config files.

This would mostly be useful to manipulate where NAV code searches for
configuration files during tests. Currently, the test suite sort of
depends on tox and the manipulation of config files below the current
virtualenv directory, but this is not conducive to running tests
interactively during development, since it would cause tests to use
the developer's currently active configuration files rather than
a set created in a temporary directory for the tests.