Support login_required in AuthorizationMiddleware
Make AuthorizationMiddleware compatible with Django's
LoginRequiredMiddlware. With the latter, if a view has the attribute
"login_required" set to False, the view is public and everyone is
authorized to access it.

The way we reuse Django's middlewares for ModPython is not compatible
with middlewares that process the view, so we need a ModPython-only
middleware for authorization.