Use term "server" instead of "endpoint" consistently in dhcpstats
This is part of simplifying the terminology used in the up-and-coming
documentation and configuration files for dhcpstats and its frontend;
terminology in the implementation and terminology used in the (user) interface
termonology should diverge as little as possible.

The documentation has more or less used the terms 'endpoint' and '(DHCP) server'
interchangeably. We fix this so that only the term '(DHCP) server' is used.
Where the user-facing interfaces are affected, backwards compability is added to
avoid the most dramatical errors.